dataset-agent/requirements.txt
smolagents[openai]>=1.24.0
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.0.0
datasets>=2.18.0
huggingface_hub>=0.24.0
kaggle>=1.6.0
selenium>=4.15.0
fake-headers>=1.0.2
tqdm>=4.66.0
python-dotenv>=1.0.0
ddgs
